move report.ts to node internal and change throws to undefined
