+++ ext/node/lib.rs
@@ -350,7 +350,6 @@ deno_core::extension!(deno_node,
     "_next_tick.ts",
     "_process/exiting.ts",
     "_process/process.ts",
-    "_process/report.ts",
     "_process/streams.mjs",
     "_readline.mjs",
     "_stream.mjs",
@@ -437,6 +436,7 @@ deno_core::extension!(deno_node,
     "internal/options.ts",
     "internal/primordials.mjs",
     "internal/process/per_thread.mjs",
+    "internal/process/report.ts",
     "internal/querystring.ts",
     "internal/readline/callbacks.mjs",
     "internal/readline/emitKeypressEvents.mjs",
