+++ ext/node/polyfills/process.ts
@@ -30,7 +30,7 @@ import {
   versions,
 } from "ext:deno_node/_process/process.ts";
 import { _exiting } from "ext:deno_node/_process/exiting.ts";
-import { report } from "ext:deno_node/_process/report.ts";
+import { report } from "ext:deno_node/internal/process/report.ts";
 export { _nextTick as nextTick, chdir, cwd, env, report, version, versions };
 import {
   createWritableStdioStream,
