+++ tools/core_import_map.json
@@ -42,7 +42,6 @@
     "ext:deno_node/_next_tick.ts": "../ext/node/polyfills/_next_tick.ts",
     "ext:deno_node/_process/exiting.ts": "../ext/node/polyfills/_process/exiting.ts",
     "ext:deno_node/_process/process.ts": "../ext/node/polyfills/_process/process.ts",
-    "ext:deno_node/_process/report.ts": "../ext/node/polyfills/_process/report.ts",
     "ext:deno_node/_process/streams.mjs": "../ext/node/polyfills/_process/streams.mjs",
     "ext:deno_node/_readline_shared_types.d.ts": "../ext/node/polyfills/_readline_shared_types.d.ts",
     "ext:deno_node/_stream.d.ts": "../ext/node/polyfills/_stream.d.ts",
@@ -139,6 +138,7 @@
     "ext:deno_node/internal/options.ts": "../ext/node/polyfills/internal/options.ts",
     "ext:deno_node/internal/primordials.mjs": "../ext/node/polyfills/internal/primordials.mjs",
     "ext:deno_node/internal/process/per_thread.mjs": "../ext/node/polyfills/internal/process/per_thread.mjs",
+    "ext:deno_node/internal/process/report.ts": "../ext/node/polyfills/internal/process/report.ts",
     "ext:deno_node/internal/querystring.ts": "../ext/node/polyfills/internal/querystring.ts",
     "ext:deno_node/internal/readline/callbacks.mjs": "../ext/node/polyfills/internal/readline/callbacks.mjs",
     "ext:deno_node/internal/readline/emitKeypressEvents.mjs": "../ext/node/polyfills/internal/readline/emitKeypressEvents.mjs",
